cicd/requirements.txt
aws-cdk.core
aws-cdk.aws_codebuild
aws-cdk.aws_codecommit
aws-cdk.aws_codepipeline
aws-cdk.aws_codepipeline_actions
aws-cdk.aws_ec2
aws-cdk.aws_events
aws-cdk.aws_events_targets
aws-cdk.aws_iam
aws-cdk.aws_lambda
aws-cdk.aws_s3
aws-cdk.aws-ecr